# docs/openapi.spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/services/{version_id}/execute": {
      "post": {
        "operationId": "ServicesController_executeWasm_v1",
        "parameters": [
          {
            "name": "version_id",
            "required": true,
            "in": "path",
            "description": "id used during upload process",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Execute a WASM bundle",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["inputs"],
                "properties": {
                  "inputs": {
                    "description": "corresponding inputs",
                    "oneOf": [
                      {
                        "type": "object",
                        "example": {
                          "a": 1,
                          "b": 2
                        }
                      },
                      {
                        "type": "array",
                        "description": "useful for synchronous batch execution using JSON data",
                        "items": {
                          "type": "object"
                        },
                        "example": [
                          {
                            "a": 1
                          },
                          {
                            "b": 2
                          },
                          {
                            "c": 3
                          }
                        ]
                      },
                      {
                        "type": "array",
                        "description": "useful for synchronous batch execution using COLUMNAR data",
                        "items": {
                          "type": "array",
                          "minItems": 2
                        },
                        "example": [
                          ["one", "two"],
                          [1, 2],
                          [3, 4]
                        ]
                      }
                    ]
                  },
                  "metadata": {
                    "description": "additional request metadata",
                    "type": "object"
                  },
                  "shared": {
                    "description": "shared data to extend every input of a batch execution",
                    "oneOf": [
                      {
                        "type": "null",
                        "example": null
                      },
                      {
                        "type": "object"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WASM executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "corresponding outputs",
                  "oneOf": [
                    {
                      "type": "object",
                      "additionalProperties": false,
                      "required": ["response_data", "response_meta"],
                      "properties": {
                        "response_data": {
                          "type": "object",
                          "properties": {
                            "outputs": {
                              "type": "object",
                              "example": {
                                "c": 3
                              }
                            },
                            "errors": {
                              "type": "array",
                              "example": []
                            },
                            "warnings": {
                              "type": "array",
                              "example": []
                            },
                            "service_chain": {
                              "type": "array",
                              "example": []
                            }
                          }
                        },
                        "response_meta": {
                          "type": "object",
                          "properties": {
                            "version_id": {
                              "type": "string"
                            },
                            "correlation_id": {
                              "type": "string"
                            },
                            "service_category": {
                              "type": "string"
                            },
                            "system": {
                              "type": "string"
                            },
                            "compiler_version": {
                              "type": "string"
                            },
                            "process_time": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "version_id": {
                          "type": "string"
                        },
                        "executed_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "duration": {
                          "type": "number"
                        },
                        "exec_times": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        },
                        "inputs": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "outputs": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "warnings": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "compiler_version": {
                          "type": "string"
                        },
                        "system": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Wrong inputs or parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["status", "message"],
                      "additionalProperties": false,
                      "properties": {
                        "status": {
                          "type": "number",
                          "description": "HTTP status code",
                          "example": 400
                        },
                        "message": {
                          "type": "string",
                          "description": "message describing the failure",
                          "example": "validation failed"
                        },
                        "cause": {
                          "type": "object",
                          "description": "description of the failure if any",
                          "example": {
                            "inputs": ["must be a valid object"]
                          },
                          "oneOf": [
                            {
                              "type": "object"
                            },
                            {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "WASM bundle not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["status", "message"],
                      "additionalProperties": false,
                      "properties": {
                        "status": {
                          "type": "number",
                          "description": "HTTP status code",
                          "example": 404
                        },
                        "message": {
                          "type": "string",
                          "description": "message describing the failure",
                          "example": "no wasm file found for version_id <id>"
                        },
                        "cause": {
                          "type": "object",
                          "description": "description of the failure if any",
                          "example": null,
                          "oneOf": [
                            {
                              "type": "object"
                            },
                            {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["status", "message"],
                      "additionalProperties": false,
                      "properties": {
                        "status": {
                          "type": "number",
                          "description": "HTTP status code",
                          "example": 413
                        },
                        "message": {
                          "type": "string",
                          "description": "message describing the failure",
                          "example": "request entity too large"
                        },
                        "cause": {
                          "type": "object",
                          "description": "description of the failure if any",
                          "example": {
                            "length": 1843,
                            "limit": 1024
                          },
                          "oneOf": [
                            {
                              "type": "object"
                            },
                            {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unable to execute WASM bundle",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["status", "message"],
                      "additionalProperties": false,
                      "properties": {
                        "status": {
                          "type": "number",
                          "description": "HTTP status code",
                          "example": 422
                        },
                        "message": {
                          "type": "string",
                          "description": "message describing the failure",
                          "example": "failed to save wasm execution history <id>"
                        },
                        "cause": {
                          "type": "object",
                          "description": "description of the failure if any",
                          "example": null,
                          "oneOf": [
                            {
                              "type": "object"
                            },
                            {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["services"]
      }
    },
    "/v1/services/{version_id}/validation": {
      "post": {
        "operationId": "ServicesController_getValidations[0]_v1",
        "parameters": [
          {
            "name": "version_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Request data and metadata",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "inputs": {
                    "description": "corresponding inputs",
                    "type": "object",
                    "example": {
                      "a": 1,
                      "b": 2
                    }
                  },
                  "metadata": {
                    "description": "additional request metadata",
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WASM validations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "WASM validations",
                  "required": ["response_data", "response_meta"],
                  "additionalProperties": false,
                  "properties": {
                    "status": {
                      "type": "string",
                      "nullable": true
                    },
                    "response_data": {
                      "type": "object",
                      "required": ["outputs", "warnings", "errors", "service_chain"],
                      "properties": {
                        "outputs": {
                          "type": "object",
                          "description": "List of validations"
                        },
                        "warnings": {
                          "type": "array",
                          "nullable": true
                        },
                        "errors": {
                          "type": "array",
                          "nullable": true
                        },
                        "service_chain": {
                          "type": "array",
                          "nullable": true
                        }
                      }
                    },
                    "response_meta": {
                      "type": "object",
                      "required": ["version_id"],
                      "properties": {
                        "version_id": {
                          "type": "string",
                          "description": "The version ID of the WASM"
                        }
                      }
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["status", "message"],
                      "additionalProperties": false,
                      "properties": {
                        "status": {
                          "type": "number",
                          "description": "HTTP status code",
                          "example": 400
                        },
                        "message": {
                          "type": "string",
                          "description": "message describing the failure",
                          "example": "validation failed"
                        },
                        "cause": {
                          "type": "object",
                          "description": "description of the failure if any",
                          "example": {
                            "inputs": ["must be a valid object"]
                          },
                          "oneOf": [
                            {
                              "type": "object"
                            },
                            {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "WASM not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["status", "message"],
                      "additionalProperties": false,
                      "properties": {
                        "status": {
                          "type": "number",
                          "description": "HTTP status code",
                          "example": 404
                        },
                        "message": {
                          "type": "string",
                          "description": "message describing the failure",
                          "example": "no wasm file found for version_id <id>"
                        },
                        "cause": {
                          "type": "object",
                          "description": "description of the failure if any",
                          "example": null,
                          "oneOf": [
                            {
                              "type": "object"
                            },
                            {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["services"]
      }
    },
...
  },
  "info": {
    "title": "wasm-service",
    "description": "API service for running WASM files",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}